Separated sender and handler management clients
